qt5: add Qml support to lupdate
Add dependency to qtdeclarative to qttools' native and nativesdk builds.
This gives us lupdate with a support for Qml files. The nativesdk builds
of qtdeclarative depend on more modules (Xml, Sql) from nativesdk-qtbase,
so preserve all libs that are already built. This brings few extra libs
to the package that might not be needed, but keeps the recipe simpler.

Signed-off-by: Samuli Piippo <samuli.piippo@theqtcompany.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>