qt5-creator: Pick native tools from native sysroot e.g. lrelease etc.
qtcreator's build system is not as creative when it comes to cross
building, it assumes to find some of tools used during translations to
be installed in bindir where its building on. In other words native
build is assumed. This patch helps the build system to locate the right
binaries during cross compilation

Fix install as we do not need chrpath'ing anymore

Fix cross-builds for non x86 systems

Signed-off-by: Khem Raj <raj.khem@gmail.com>